requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
pandas>=2.0
google-generativeai>=0.7
python-dotenv>=1.0
plotly>=5.17   # 仅 --plot 时需要
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
jieba>=0.42.1
tqdm>=4.65.0 